Automates cloud-init for a SQL server
TL;DR
-----

Generates a cloud-init for OS database with MSSQL

Details
-------

Creates a YAML tempplate (using `ytt`) to generate user data to
customize a Focal server for use as a SQL Server for OutSystems.
Includes a script `generate-cloud-init` that generates the user
data for encoding and setting as an OVF property.