Fixes `.envrc` when params haven't been decrypted (#3)
TL;DR
-----

Repairs a defect where `direnv` would error out if params.yml
did not exist

Details
-------

The value for `GOOGLE_CREDENTIALS` is set based on one of the
parameters in `params.yml`. This caused an error when either
`params.yml` had not been created or was still encrypted. This
change fixes that error.